refactor(stats): split up weekly pulse functions
