+++ projects/client/src/lib/sections/stats/_internal/models/DateRange.ts
@@ -1,0 +1,1 @@
+export type DateRange = { readonly start: Date; readonly end: Date };
