+++ projects/client/src/lib/sections/stats/_internal/models/PulseGraphItem.ts
@@ -1,4 +1,5 @@
-import type { PulseGraphData, PulseGraphType } from '../pulseGraphs.ts';
+import type { PulseGraphData } from './PulseGraphData.ts';
+import type { PulseGraphType } from './PulseGraphType.ts';
 
 export type PulseGraphItem = {
   readonly type: 'graph';
