+++ projects/client/src/lib/sections/stats/_internal/models/PulseGraphType.ts
@@ -1,0 +1,6 @@
+export type PulseGraphType =
+  | 'dailyBars'
+  | 'weekTrend'
+  | 'watchClock'
+  | 'showsMovies'
+  | 'ratingsDistribution';
