+++ projects/client/src/lib/sections/stats/_internal/models/PulseStat.ts
@@ -1,0 +1,9 @@
+export type PulseStat = {
+  readonly key: string;
+  readonly rawValue: number;
+  readonly value: string;
+  readonly label: string;
+  readonly tooltip?: string;
+  readonly delta: number | null;
+  readonly note?: string;
+};
