+++ projects/client/src/lib/sections/stats/_internal/models/WeekData.ts
@@ -1,0 +1,6 @@
+export type WeekData = {
+  readonly movieDates: readonly Date[];
+  readonly showDates: readonly Date[];
+  readonly uniqueShows: number;
+  readonly ratings: ReadonlyArray<{ readonly rating: number }>;
+};
