+++ projects/client/src/lib/sections/stats/_internal/rankStats.ts
@@ -1,58 +1,4 @@
-import { getDayKey } from '$lib/utils/date/getDayKey.ts';
-
-export type PulseStat = {
-  readonly key: string;
-  readonly rawValue: number;
-  readonly value: string;
-  readonly label: string;
-  readonly tooltip?: string;
-  readonly delta: number | null;
-  readonly note?: string;
-};
-
-export function countUniqueDays(dates: ReadonlyArray<Date>): number {
-  return new Set(dates.map(getDayKey)).size;
-}
-
-export function getBusiestDay(
-  dates: ReadonlyArray<Date>,
-): { readonly dayIndex: number; readonly count: number } | null {
-  if (dates.length === 0) return null;
-
-  const counts = dates.reduce((acc, d) => {
-    const dow = d.getDay();
-    acc.set(dow, (acc.get(dow) ?? 0) + 1);
-    return acc;
-  }, new Map<number, number>());
-
-  return [...counts.entries()].reduce(
-    (best, [day, count]) =>
-      count > best.count ? { dayIndex: day, count } : best,
-    { dayIndex: 0, count: 0 },
-  );
-}
-
-export function dayOfWeekDate(dayIndex: number, now: Date): Date {
-  const date = new Date(now);
-  date.setDate(date.getDate() - ((date.getDay() - dayIndex + 7) % 7));
-  return date;
-}
-
-export function maxPlaysInSingleDay(dates: ReadonlyArray<Date>): number {
-  if (dates.length === 0) return 0;
-
-  const counts = dates.reduce((acc, d) => {
-    const key = getDayKey(d);
-    acc.set(key, (acc.get(key) ?? 0) + 1);
-    return acc;
-  }, new Map<string, number>());
-
-  return Math.max(...counts.values());
-}
-
-export function computeDelta(thisWeek: number, lastWeek: number): number {
-  return thisWeek - lastWeek;
-}
+import type { PulseStat } from './models/PulseStat.ts';
 
 const scoreWeights = {
   delta: 4,
