+++ projects/client/src/lib/sections/stats/_internal/utils/computeDelta.ts
@@ -1,0 +1,3 @@
+export function computeDelta(thisWeek: number, lastWeek: number): number {
+  return thisWeek - lastWeek;
+}
