+++ projects/client/src/lib/sections/stats/_internal/utils/countUniqueDays.ts
@@ -1,0 +1,5 @@
+import { getDayKey } from '$lib/utils/date/getDayKey.ts';
+
+export function countUniqueDays(dates: ReadonlyArray<Date>): number {
+  return new Set(dates.map(getDayKey)).size;
+}
