+++ projects/client/src/lib/sections/stats/_internal/utils/normalizeScore.ts
@@ -1,0 +1,7 @@
+export function normalizeScore(
+  raw: number,
+  maxPossible: number,
+): number {
+  if (maxPossible <= 0) return 0;
+  return Math.min((raw / maxPossible) * 100, 100);
+}
